Move block-encoding arithmetic into block_encoding_combination
Splits the composite encodings and the arithmetic that builds them out of
block_encoding_base.py, which had grown to just under 2000 lines and is
about to take a ProductBlockEncoding as well.

The arithmetic moves with the classes rather than staying behind, because
that is what keeps the package acyclic. BlockEncoding has to be able to
construct a LinearCombinationBlockEncoding, and LinearCombinationBlockEncoding
has to subclass BlockEncoding; keeping both in the new module means the
dependency runs one way only, from block_encoding_combination to
block_encoding_base. The operators are attached to BlockEncoding in
block_encoding.py, exactly as the constructors and transformations already
are, and re-declared under TYPE_CHECKING so static analysis still sees them
as ordinary members. Special methods are looked up on the type, so assigning
them is equivalent to defining them in the class body.

block_encoding_base.py now imports nothing from its own package.

Also redirects transformations/pseudo_inv.py to import BlockEncoding from
block_encoding_base rather than from block_encoding. It was the only
transformation not doing so, and it formed the package's one genuine import
cycle (block_encoding -> transformations -> pseudo_inv -> block_encoding),
which worked only because BlockEncoding happened to be bound before the
transformations import ran.

The moved code is byte-identical to the original apart from a four-space
dedent, the renames to the build_*/apply_* convention used by the sibling
modules, the dropped @classmethod decorators, and the __rmul__ alias now
applied by block_encoding.py.
