+++ src/qrisp/block_encodings/block_encoding_base.py
@@ -20,21 +20,17 @@ from __future__ import annotations
 
 from collections.abc import Callable, Sequence
 from dataclasses import dataclass
-from types import NotImplementedType
 from typing import TYPE_CHECKING, Any
 
 import jax
 import jax.numpy as jnp
 import numpy as np
-from jax.errors import TracerArrayConversionError
 from jax.tree_util import register_pytree_node_class
 from jax.typing import ArrayLike
 
 from qrisp.alg_primitives.reflection import reflection
-from qrisp.alg_primitives.state_preparation import prepare
-from qrisp.block_encodings.ancilla_layout import _AncillaLayout, _maximum_layout_size
 from qrisp.core import QuantumVariable
-from qrisp.core.gate_application_functions import gphase, h, measure, reset, ry, x, z
+from qrisp.core.gate_application_functions import h, measure, reset, x
 from qrisp.environments import conjugate, control, invert
 from qrisp.jasp import (
     RUS,
@@ -43,14 +39,12 @@ from qrisp.jasp import (
     expectation_value,
     jrange,
     num_qubits,
-    q_switch,
-    qache,
 )
 from qrisp.jasp import (
     check_for_tracing_mode as is_tracing,
 )
 from qrisp.jasp.tracing_logic import QuantumVariableTemplate
-from qrisp.qtypes import QuantumBool, QuantumFloat
+from qrisp.qtypes import QuantumBool
 
 if TYPE_CHECKING:
     from qrisp.interface.backend import BackendLike
@@ -59,14 +53,6 @@ if TYPE_CHECKING:
     # tree_flatten/tree_unflatten below.
     PyTreeChildren = tuple[ArrayLike, list[QuantumVariableTemplate]]
     PyTreeAuxData = tuple[Callable[..., None], int, bool]
-
-
-def _is_non_negative_real(value: Any) -> bool:
-    try:
-        value = np.asarray(value)
-    except Exception:
-        return False
-    return bool(np.all(np.isreal(value)) and np.all(np.real(value) >= 0))
 
 
 @register_pytree_node_class
@@ -1051,548 +1037,33 @@ class BlockEncoding:
     def _get_lcu_terms(self) -> _LCUTerms:
         return ((1, self),)
 
-    @classmethod
-    def linear_combination(
-        cls,
-        block_encodings: list[BlockEncoding],
-        coefficients: "ArrayLike | None" = None,
-    ) -> BlockEncoding:
-        r"""Returns a BlockEncoding of a linear combination of operators.
-
-        This method implements the linear combination $\sum_i\alpha_iA_i$ via the LCU
-        (Linear Combination of Unitaries) framework, where $A_i$ are
-        the operators encoded by the respective instances, and $\alpha_i$ are real coefficients.
-
-        Equivalently, block-encoded operators can also be combined via ``+``, ``-`` and scalar multiplication.
-
-        Parameters
-        ----------
-        block_encodings : list[BlockEncoding]
-            Block-encodings of operators acting on the same operands.
-        coefficients : ArrayLike, optional
-            Coefficients multiplying the encoded operators. Defaults to one
-            for every block-encoding.
-
-        Returns
-        -------
-        BlockEncoding
-            A block-encoding of the linear combination.
-
-        Raises
-        ------
-        ValueError
-            If no block-encodings are supplied, the coefficient count does
-            not match, or operand counts differ.
-        TypeError
-            If an item is not a BlockEncoding.
-
-        """
-        if len(block_encodings) == 0:
-            raise ValueError("At least one block-encoding is required.")
-
-        if coefficients is None:
-            coefficients = [1] * len(block_encodings)
-        elif len(coefficients) != len(block_encodings):
-            raise ValueError("The number of coefficients must match the number of block-encodings.")
-
-        terms = []
-        for block_encoding, coefficient in zip(block_encodings, coefficients):
-            if not isinstance(block_encoding, BlockEncoding):
-                raise TypeError(f"Expected every item to be a BlockEncoding, but got {type(block_encoding).__name__}.")
-            terms.append((coefficient, block_encoding))
-
-        return cls._from_lcu_terms(terms)
-
-    @classmethod
-    def _from_lcu_terms(cls, terms: Sequence[_LCUTerm]) -> BlockEncoding:
-        """Build a linear-combination block encoding from weighted terms."""
-        terms = _canonicalize_lcu_terms(terms)
-        if len(terms) == 0:
-            raise ValueError("Cannot construct a block encoding from an all-zero linear combination.")
-        if len(terms) == 1:
-            coefficient, block_encoding = terms[0]
-            if isinstance(coefficient, (int, float, complex, np.number)) and coefficient == 1:
-                return block_encoding
-        return LinearCombinationBlockEncoding(terms)
-
-    def __add__(self, other: BlockEncoding) -> BlockEncoding:
-        r"""Returns a BlockEncoding of the sum of two operators.
-
-        This method implements the addition $A + B$ via the LCU
-        (Linear Combination of Unitaries) framework, where $A$ and $B$ are
-        the operators encoded by the respective instances.
-
-        Parameters
-        ----------
-        other : BlockEncoding
-            The BlockEncoding instance to be added.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding instance representing the operator sum.
-
-        Notes
-        -----
-        - Can only be used when both BlockEncodings have the same operand structure.
-        - The ``+`` operator should be used sparingly, primarily to combine a few block encodings.
-          For larger-scale polynomial transformations,
-          Quantum Signal Processing (QSP) is the superior method.
-
-        Examples
-        --------
-        Define two block-encodings and add them.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            H1 = X(0)*X(1) + 0.2*Y(0)*Y(1)
-            H2 = Z(0)*Z(1) + X(2)
-            H3 = H1 + H2
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = BlockEncoding.from_operator(H3)
-
-            BE_add = BE1 + BE2
-
-            def operand_prep():
-                qv = QuantumFloat(3)
-                return qv
-
-            @terminal_sampling
-            def main(BE):
-                qv = BE.apply_rus(operand_prep)()
-                return qv
-
-            res_be3 = main(BE3)
-            res_be_add = main(BE_add)
-            print("Result from BE of H1 + H2: ", res_be3)
-            print("Result from BE1 + BE2: ", res_be_add)
-            # Result from BE of H1 + H2:  {0: 0.37878788804466035, 4: 0.37878788804466035, 3: 0.24242422391067933}
-            # Result from BE1 + BE2:  {0: 0.37878789933341894, 4: 0.37878789933341894, 3: 0.24242420133316217}
-
-        """
-        if not isinstance(other, BlockEncoding):
-            return NotImplemented
-
-        return type(self)._from_lcu_terms(self._get_lcu_terms() + other._get_lcu_terms())
-
-    def __sub__(self, other: BlockEncoding) -> BlockEncoding:
-        r"""Returns a BlockEncoding of the difference between two operators.
-
-        This method implements the subtraction $A - B$ via the LCU
-        (Linear Combination of Unitaries) framework, where $A$ and $B$ are
-        the operators encoded by the respective instances.
-
-        Parameters
-        ----------
-        other : BlockEncoding
-            The BlockEncoding instance to be subtracted.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding representing the operator difference.
-
-        Notes
-        -----
-        - Can only be used when both BlockEncodings have the same operand structure.
-        - The ``-`` operator should be used sparingly, primarily to combine a few block encodings.
-          For larger-scale polynomial transformations,
-          Quantum Signal Processing (QSP) is the superior method.
-
-        Examples
-        --------
-        Define two block-encodings and subtract them.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            H1 = X(0)*X(1) + 0.2*Y(0)*Y(1)
-            H2 = Z(0)*Z(1) + X(2)
-            H3 = H1 - H2
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = BlockEncoding.from_operator(H3)
-
-            BE_sub = BE1 - BE2
-
-            def operand_prep():
-                qv = QuantumFloat(3)
-                return qv
-
-            @terminal_sampling
-            def main(BE):
-                qv = BE.apply_rus(operand_prep)()
-                return qv
-
-            res_be3 = main(BE3)
-            res_be_sub = main(BE_sub)
-            print("Result from BE of H1 - H2: ", res_be3)
-            print("Result from BE1 - BE2: ", res_be_sub)
-            # Result from BE of H1 - H2:  {0: 0.37878788804466035, 4: 0.37878788804466035, 3: 0.24242422391067933}
-            # Result from BE1 - BE2:  {0: 0.37878789933341894, 4: 0.37878789933341894, 3: 0.24242420133316217}
-
-        """
-        if not isinstance(other, BlockEncoding):
-            return NotImplemented
-
-        other_terms = tuple((-coefficient, block_encoding) for coefficient, block_encoding in other._get_lcu_terms())
-        return type(self)._from_lcu_terms(self._get_lcu_terms() + other_terms)
-
-    def __mul__(self, other: "ArrayLike") -> BlockEncoding:
-        r"""Returns a BlockEncoding of the scaled operator.
-
-        This method implements the scalar multiplication $c \cdot A$, where $A$
-        is the operator encoded by this instance and $c$ is the
-        provided scalar.
-
-        Parameters
-        ----------
-        other : ArrayLike
-            The scalar scaling factor (coefficient) to apply. Can be a Python float,
-            a JAX/NumPy scalar, or a 0-dimensional array.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding instance representing the scaled operator.
-
-        Notes
-        -----
-        - Multiplying by a scalar $c$ results in a new BlockEncoding of $cA$ by updating $\alpha \rightarrow c\alpha$.
-
-        Examples
-        --------
-        Define two block-encodings and implement their scaled sum as a new block encoding.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            # Commuting operators H1 and H2
-            H1 = X(0)*X(1) + 0.2*Y(0)*Y(1)
-            H2 = Z(0)*Z(1) + X(2)
-            H3 = 2*H1 + H2
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = BlockEncoding.from_operator(H3)
-
-            BE_mul = 2*BE1 + BE2
-            BE_mul_r = BE1*2 + BE2
-
-            def operand_prep():
-                qv = QuantumFloat(3)
-                return qv
-
-            @terminal_sampling
-            def main(BE):
-                qv = BE.apply_rus(operand_prep)()
-                return qv
-
-            res_be3 = main(BE3)
-            res_be_mul = main(BE_mul)
-            res_be_mul_r = main(BE_mul_r)
-
-            print("Result from BE of 2 * H1 + H2: ", res_be3)
-            print("Result from 2 * BE1 + BE2: ", res_be_mul)
-            print("Result from BE1 * 2 + BE2: ", res_be_mul_r)
-            # Result from BE of 2 * H1 + H2:  {3.0: 0.5614033770142979, 0.0: 0.21929831149285103, 4.0: 0.21929831149285103}
-            # Result from 2 * BE1 + BE2:  {3.0: 0.5614033770142979, 0.0: 0.21929831149285103, 4.0: 0.21929831149285103}
-            # Result from BE1 * 2 + BE2:  {3.0: 0.5614033770142979, 0.0: 0.21929831149285103, 4.0: 0.21929831149285103}
-
-        """
-
-        if isinstance(other, ArrayLike):
-            terms = [(other * coefficient, block_encoding) for coefficient, block_encoding in self._get_lcu_terms()]
-            return type(self)._from_lcu_terms(terms)
-
-        return NotImplemented
-
-    def __matmul__(self, other: BlockEncoding) -> BlockEncoding:
-        r"""Returns a BlockEncoding of the product of two operators.
-
-        This method implements the operator product $A \cdot B$ by composing
-        two BlockEncodings, where $A$ and $B$ are the operators encoded by the respective instances.
-
-        Parameters
-        ----------
-        other : BlockEncoding
-            The BlockEncoding instance to be multiplied.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding representing the operator product.
-
-        Notes
-        -----
-        - Can only be used when both BlockEncodings have the same operand structure.
-        - The ``@`` operator should be used sparingly, primarily to combine a few block encodings. For larger-scale polynomial transformations, Quantum Signal Processing (QSP) is the superior method.
-        - The product of two Hermitian operators A and B is Hermitian if and only if they commute, i.e., AB = BA.
-
-        Examples
-        --------
-        Define two block-encodings and multiply them.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            # Commuting operators H1 and H2
-            H1 = X(0)*X(1) + 0.2*Y(0)*Y(1)
-            H2 = Z(0)*Z(1) + X(2)
-            H3 = H1 * H2
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = BlockEncoding.from_operator(H3)
-
-            BE_mul = BE1 @ BE2
-
-            def operand_prep():
-                qv = QuantumFloat(3)
-                return qv
-
-            @terminal_sampling
-            def main(BE):
-                qv = BE.apply_rus(operand_prep)()
-                return qv
-
-            res_be3 = main(BE3)
-            res_be_mul = main(BE_mul)
-            print("Result from BE of H1 * H2: ", res_be3)
-            print("Result from BE1 @ BE2: ", res_be_mul)
-            # Result from BE of H1 * H2:  {3.0: 0.5, 7.0: 0.5}
-            # Result from BE1 @ BE2:  {3.0: 0.5, 7.0: 0.5}
-
-        """
-        if not isinstance(other, BlockEncoding):
-            return NotImplemented
-
-        m = len(self._anc_templates)
-        n = len(other._anc_templates)
-
-        def new_unitary(*args):
-            other_args = args[m : m + n] + args[m + n :]
-            other.unitary(*other_args)
-            self_args = args[:m] + args[m + n :]
-            self.unitary(*self_args)
-
-        new_anc_templates = self._anc_templates + other._anc_templates
-        new_alpha = self.alpha * other.alpha
-        return BlockEncoding(new_alpha, new_anc_templates, new_unitary, num_ops=self.num_ops)
-
-    def __radd__(self, other: Any) -> BlockEncoding | NotImplementedType:
-        """Support adding a BlockEncoding to the start value used by sum()."""
-        if other == 0:
-            return self
-        return NotImplemented
-
-    __rmul__ = __mul__
-
-    def kron(self, other: BlockEncoding) -> BlockEncoding:
-        r"""Returns a BlockEncoding of the Kronecker product (tensor product) of two operators.
-
-        This method implements the operator $A \otimes B$, where $A$ and $B$ are
-        the operators encoded by the respective instances. Following the
-        construction in Chapter 10.2 in `Dalzell et al. <https://arxiv.org/abs/2310.03011>`_,
-        the resulting BlockEncoding is formed by the tensor product of the underlying unitaries, $U_A \otimes U_B$.
-
-        Parameters
-        ----------
-        other : BlockEncoding
-            The BlockEncoding instance to be tensored.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding representing the tensor product $A \otimes B$.
-
-        Notes
-        -----
-        - **Normalization**: The normalization factors ($\alpha$) are combined multiplicatively.
-        - The ``kron`` operator maps the operands of self to the first set of operands and the operands of other to the remaining operands in a single unified unitary.
-        - The ``kron`` operator should be used sparingly, primarily to combine a few block encodings.
-        - A more qubit-efficient implementation of the Kronecker product can be found in `this paper <https://arxiv.org/pdf/2509.15779>`_ and will be implemented in future updates.
-
-        Examples
-        --------
-        **Example 1:**
-
-        Define two block-encodings and perform their Kronecker product.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            H1 = X(0)*X(1) + 0.2*Y(0)*Y(1)
-            H2 = Z(0)*Z(1) + X(2)
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-
-            BE_composed = BE1.kron(BE2)
-
-            n1 = H1.find_minimal_qubit_amount()
-            n2 = H2.find_minimal_qubit_amount()
-
-            def operand_prep():
-                qv1 = QuantumVariable(n1)
-                qv2 = QuantumVariable(n2)
-                return qv1, qv2
-
-            @terminal_sampling
-            def main(BE):
-                return BE.apply_rus(operand_prep)()
-
-            result = main(BE_composed)
-            print("Result from BE1.kron(BE2): ", result)
-
-        **Example 2:**
-
-        Perform multiple Kronecker products of block-encodings in sequence.
-
-        ::
-
-            from qrisp import *
-            from qrisp.operators import X, Y, Z
-
-            H1 = X(0)*X(1)
-            H2 = Z(0)*Z(1)
-            H3 = Y(0)*Y(1)
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = BlockEncoding.from_operator(H3)
-
-            # Compose BE1 with the composition of BE2 and BE3
-            BE_composed = BE1.kron(BE2.kron(BE3))
-
-            n1 = H1.find_minimal_qubit_amount()
-            n2 = H2.find_minimal_qubit_amount()
-            n3 = H3.find_minimal_qubit_amount()
-
-            def operand_prep():
-                qv1 = QuantumVariable(n1)
-                qv2 = QuantumVariable(n2)
-                qv3 = QuantumVariable(n3)
-                return qv1, qv2, qv3
-
-            @terminal_sampling
-            def main(BE):
-                return BE.apply_rus(operand_prep)()
-
-            result = main(BE_composed)
-            print("Result from BE1.kron(BE2.kron(BE3)): ", result)
-
-        """
-        m = len(self._anc_templates)
-        n = len(other._anc_templates)
-
-        def new_unitary(*args):
-            self_ancs = args[:m]
-            other_ancs = args[m : m + n]
-            operands = args[m + n :]
-
-            self.unitary(*self_ancs, *operands[: self.num_ops])
-            other.unitary(*other_ancs, *operands[self.num_ops :])
-
-        new_anc_templates = self._anc_templates + other._anc_templates
-        new_alpha = self.alpha * other.alpha
-        return BlockEncoding(
-            new_alpha,
-            new_anc_templates,
-            new_unitary,
-            num_ops=self.num_ops + other.num_ops,
-            is_hermitian=self.is_hermitian and other.is_hermitian,
-        )
-
-    def __neg__(self) -> BlockEncoding:
-        r"""Returns a BlockEncoding of the negated operator.
-
-        This method implements the transformation $A \to -A$ by scaling the
-        encoded operator by $-1$.
-
-        Returns
-        -------
-        BlockEncoding
-            A new BlockEncoding instance representing the operator $-A$.
-
-        Examples
-        --------
-        Define a block-encoding and negate it.
-
-        ::
-
-            from qrisp import *
-            from qrisp.block_encodings import BlockEncoding
-            from qrisp.operators import X, Y, Z
-
-            H1 = X(0)*X(1) - 0.2*Y(0)*Y(1)
-            H2 = 0.2*Y(0)*Y(1) - X(0)*X(1)
-
-            BE1 = BlockEncoding.from_operator(H1)
-            BE2 = BlockEncoding.from_operator(H2)
-            BE3 = -BE1
-
-            def operand_prep():
-                qv = QuantumFloat(3)
-                return qv
-
-            @terminal_sampling
-            def main(BE):
-                qv = BE.apply_rus(operand_prep)()
-                return qv
-
-            res_be2 = main(BE2)
-            res_be_neg = main(BE3)
-
-            print("Result from BE of H2 = - H1: ", res_be2)
-            print("Result from - BE1: ", res_be_neg)
-            # Result from BE of H2 = - H1:  {3.0: 1.0}
-            # Result from - BE1:  {3.0: 1.0}
-
-        """
-
-        def new_unitary(*args):
-            self.unitary(*args)
-            gphase(np.pi, args[0][0])
-
-        return BlockEncoding(
-            self.alpha,
-            self._anc_templates,
-            new_unitary,
-            num_ops=self.num_ops,
-            is_hermitian=self.is_hermitian,
-        )
-
     # ------------------------------------------------------------------
     # The methods below are attached to this class after its definition, in
     # block_encoding.py: each one is implemented in its own module under
-    # constructors/ or transformations/, and each of those modules needs to
-    # import BlockEncoding itself, so importing them here at runtime would
-    # be circular. Re-declaring them under TYPE_CHECKING (never executed at
-    # runtime) makes them visible to type checkers as ordinary members of
-    # this class, without changing any runtime behaviour or reintroducing
-    # that circular import.
+    # constructors/ or transformations/, or in block_encoding_combination.py,
+    # and each of those modules needs to import BlockEncoding itself, so
+    # importing them here at runtime would be circular. Re-declaring them
+    # under TYPE_CHECKING (never executed at runtime) makes them visible to
+    # type checkers as ordinary members of this class, without changing any
+    # runtime behaviour or reintroducing that circular import.
+    #
+    # The arithmetic operators live in block_encoding_combination.py because
+    # they construct the composite encodings defined there, which in turn
+    # subclass this class. Keeping them there is what makes that dependency
+    # run one way only.
     # ------------------------------------------------------------------
     if TYPE_CHECKING:
+        from .block_encoding_combination import (
+            apply_add,
+            apply_kron,
+            apply_matmul,
+            apply_mul,
+            apply_neg,
+            apply_radd,
+            apply_sub,
+            build_from_lcu_terms,
+            build_linear_combination,
+        )
         from .constructors import (
             build_from_array,
             build_from_eye,
@@ -1624,364 +1095,18 @@ class BlockEncoding:
         sim = apply_sim
         svt = apply_svt
 
+        linear_combination = classmethod(build_linear_combination)
+        _from_lcu_terms = classmethod(build_from_lcu_terms)
+
+        __add__ = apply_add
+        __radd__ = apply_radd
+        __sub__ = apply_sub
+        __mul__ = apply_mul
+        __rmul__ = apply_mul
+        __matmul__ = apply_matmul
+        __neg__ = apply_neg
+        kron = apply_kron
+
 
 _LCUTerm = tuple[ArrayLike, BlockEncoding]
 _LCUTerms = tuple[_LCUTerm, ...]
-
-
-def _is_statically_zero(value: Any) -> bool:
-    """Return whether ``value`` is a concrete scalar zero."""
-    if isinstance(value, jax.core.Tracer):
-        return False
-    try:
-        value = np.asarray(value)
-    except Exception:
-        return False
-    return value.ndim == 0 and bool(value == 0)
-
-
-def _canonicalize_lcu_terms(terms: Sequence[_LCUTerm]) -> _LCUTerms:
-    """Merge identity-equal child encodings and remove concrete zero terms."""
-    merged_terms: list[_LCUTerm] = []
-    for coefficient, block_encoding in terms:
-        for index, (_, existing_block_encoding) in enumerate(merged_terms):
-            if existing_block_encoding is block_encoding:
-                merged_terms[index] = (merged_terms[index][0] + coefficient, block_encoding)
-                break
-        else:
-            merged_terms.append((coefficient, block_encoding))
-
-    return tuple(
-        (coefficient, block_encoding)
-        for coefficient, block_encoding in merged_terms
-        if not _is_statically_zero(coefficient)
-    )
-
-
-def _is_reusable_unitary(block_encoding: BlockEncoding) -> bool:
-    """Return whether ``block_encoding.unitary`` yields the same object on every access.
-
-    A plain BlockEncoding stores its unitary in a field, so it is always reusable.
-    A LinearCombinationBlockEncoding derives its unitary and only caches it when the
-    derivation captured no JAX tracer. Reusing a child unitary that is rebuilt on
-    every access would hand every enclosing ``q_switch`` a fresh Python closure,
-    which is exactly what this module's caching exists to avoid.
-
-    The child's ``unitary`` must have been accessed before this is called, so that
-    a cacheable child has had the chance to populate its cache.
-    """
-    if isinstance(block_encoding, LinearCombinationBlockEncoding):
-        return "_cached_unitary" in block_encoding.__dict__
-    return True
-
-
-def _make_lcu_branch(
-    child_unitary: Callable[..., None],
-    layout: _AncillaLayout,
-    name: str,
-) -> Callable[..., None]:
-    """Build one SELECT branch applying ``child_unitary`` to views into the workspace.
-
-    The shared workspace is taken as the branch's first operand rather than being
-    captured from the enclosing scope, so the branch can be built once, ahead of
-    any tracing, and reused for every invocation of the LCU unitary.
-    """
-
-    def branch(shared_ancilla, *operands):
-        child_unitary(*layout.construct_views(shared_ancilla), *operands)
-
-    branch.__name__ = name
-    # Caching the branch body is what keeps the repeated tracing of q_switch cheap:
-    # the tree-based q_switch traces every branch several times (for the loop body,
-    # for both arms of its final conditional) and custom_control/custom_inversion
-    # speculatively trace the controlled and inverted variants on top of that. With
-    # a qached body all but the first of those become pjit cache hits.
-    return qache(branch)
-
-
-def _identity_lcu_branch(shared_ancilla, *operands) -> None:
-    """Pad the SELECT to a power of two; selected only for zero-amplitude indices."""
-
-
-@register_pytree_node_class
-class LinearCombinationBlockEncoding(BlockEncoding):
-    """A block encoding represented by an immutable tuple of LCU terms."""
-
-    def __init__(self, terms: Sequence[_LCUTerm]) -> None:
-        r"""Initialize a linear-combination block encoding from weighted terms.
-
-        Each term is a ``(coefficient, block_encoding)`` pair. If the child
-        block-encoding represents $A_i / \alpha_i$, the combined LCU uses
-        ``coefficient * child.alpha`` as the coefficient of its child unitary.
-        The resulting normalization is therefore
-        ``sum(abs(coefficient * child.alpha))``.
-
-        The child ancillas are packed into one shared workspace after one
-        selector register. The workspace is sized to the largest child layout,
-        and each selected branch reconstructs its typed ancillas as views into
-        that workspace. Since the selector chooses exactly one child unitary,
-        nested sums are lowered to one preparation/select/preparation sequence.
-        The terms are retained as the immutable authoritative representation;
-        the compiled unitary and metadata are derived from them.
-
-        Parameters
-        ----------
-        terms : tuple[tuple[ArrayLike, BlockEncoding], ...]
-            Weighted block-encoding terms. The terms must be non-empty, and all
-            child block-encodings must have the same number of operands.
-
-        Raises
-        ------
-        ValueError
-            If no terms are supplied or the child operand counts differ.
-        TypeError
-            If a term does not contain a BlockEncoding.
-
-        Notes
-        -----
-        A single term with coefficient ``1`` is returned as the original child
-        by the factory that constructs this class. A single term with another
-        coefficient uses the child's ancillas directly and applies the required
-        phase for a negative coefficient.
-
-        Everything derived from the terms — the unitary, its SELECT branches, the
-        ancilla layouts and templates, the coefficients and the normalization — is
-        built once and cached, and derived quantities that are known at build time
-        are kept as NumPy values rather than JAX arrays. Both are required for the
-        compilation cost to stay proportional to the size of the expression: Jasp's
-        caches are keyed on object identity, and ``prepare`` selects its algorithm
-        by whether its amplitude vector converts to NumPy. Derived values that
-        capture a JAX tracer are not cached, since reusing them in a later trace
-        would leak the tracer.
-
-        """
-        terms = _canonicalize_lcu_terms(terms)
-        if len(terms) == 0:
-            raise ValueError("Cannot construct a block encoding from an all-zero linear combination.")
-        if any(not isinstance(block_encoding, BlockEncoding) for _, block_encoding in terms):
-            raise TypeError("Expected every item to be a BlockEncoding.")
-
-        num_ops = terms[0][1].num_ops
-        if any(block_encoding.num_ops != num_ops for _, block_encoding in terms):
-            raise ValueError("All block-encodings must have the same number of operands.")
-
-        self._terms = terms
-
-    def __setattr__(self, name: str, value: Any) -> None:
-        """Prevent reassignment of the authoritative term representation."""
-        if name == "_terms" and hasattr(self, "_terms"):
-            raise AttributeError("Linear-combination terms are immutable.")
-        object.__setattr__(self, name, value)
-
-    @property
-    def terms(self) -> _LCUTerms:
-        """The immutable weighted child block encodings."""
-        return self._terms
-
-    #
-    # Derived state
-    #
-    # Everything below is derived from the authoritative terms. It is cached on
-    # first access, because rebuilding it would produce fresh Python objects every
-    # time, and Jasp's caches (qache, and hence jax.jit) are keyed on object
-    # identity. Without the cache, every enclosing q_switch re-traces the whole
-    # subtree from scratch, and nested linear combinations multiply that cost once
-    # per nesting level.
-    #
-    # A derived value is only cached when it is independent of any JAX trace.
-    # Values that capture tracers must be rebuilt per trace: a tracer leaked into
-    # a later trace raises a JAX leak error.
-    #
-
-    @property
-    def _lcu_selector_size(self) -> int:
-        """Return the number of qubits addressing the SELECT branches."""
-        return (len(self.terms) - 1).bit_length()
-
-    @property
-    def _lcu_layouts(self) -> tuple[_AncillaLayout, ...]:
-        """Return the packing of every child's ancillas into the shared workspace."""
-        cached = self.__dict__.get("_cached_layouts")
-        if cached is not None:
-            return cached
-
-        layouts = tuple(
-            _AncillaLayout.from_templates(block_encoding._anc_templates) for _, block_encoding in self.terms
-        )
-        if all(layout.has_static_sizes for layout in layouts):
-            object.__setattr__(self, "_cached_layouts", layouts)
-        return layouts
-
-    @property
-    def _lcu_coefficients(self) -> ArrayLike:
-        r"""Return $\alpha_i c_i$ for every term, i.e. the LCU coefficients of the child unitaries.
-
-        The result is a concrete NumPy array whenever the coefficients and the child
-        normalizations are known at build time. Staying out of ``jnp`` here is not a
-        micro-optimization: inside a JAX trace even ``jnp.abs`` of a compile-time
-        constant returns a tracer, and :func:`qrisp.prepare` dispatches on whether its
-        amplitude vector converts to NumPy. A traced amplitude vector silently
-        downgrades the state preparation from the single-gate ``prepare_qiskit`` path
-        to the fully symbolic ``prepare_qswitch`` path, which costs orders of magnitude
-        more equations and is applied twice (PREP and its inverse).
-
-        The coefficients are genuinely traced when the block encoding crossed a jit
-        boundary as a pytree, in which case the symbolic path is the correct one.
-        """
-        cached = self.__dict__.get("_cached_coefficients")
-        if cached is not None:
-            return cached
-
-        try:
-            coefficients = np.array(
-                [
-                    np.asarray(coefficient) * np.asarray(block_encoding.alpha)
-                    for coefficient, block_encoding in self.terms
-                ],
-                dtype=complex,
-            )
-        except TracerArrayConversionError:
-            return jnp.array(
-                [coefficient * block_encoding.alpha for coefficient, block_encoding in self.terms],
-                dtype=complex,
-            )
-
-        object.__setattr__(self, "_cached_coefficients", coefficients)
-        return coefficients
-
-    @property
-    def _lcu_amplitudes(self) -> ArrayLike:
-        """Return the PREP amplitudes, padded to the selector dimension."""
-        coefficients = self._lcu_coefficients
-        xp = np if isinstance(coefficients, np.ndarray) else jnp
-
-        padded = xp.pad(coefficients, (0, (1 << self._lcu_selector_size) - len(self.terms)))
-        return xp.sqrt(padded / xp.sum(xp.abs(padded)))
-
-    @property
-    def alpha(self) -> ArrayLike:
-        """Return the normalization derived from the weighted child encodings."""
-        coefficients = self._lcu_coefficients
-        xp = np if isinstance(coefficients, np.ndarray) else jnp
-        return xp.sum(xp.abs(coefficients))
-
-    @property
-    def _anc_templates(self) -> list[QuantumVariableTemplate]:
-        cached = self.__dict__.get("_cached_anc_templates")
-        if cached is not None:
-            return list(cached)
-
-        if len(self.terms) == 1:
-            templates = tuple(self.terms[0][1]._anc_templates)
-            object.__setattr__(self, "_cached_anc_templates", templates)
-            return list(templates)
-
-        layouts = self._lcu_layouts
-        templates = (
-            QuantumFloat(self._lcu_selector_size).template(),
-            QuantumVariable(_maximum_layout_size(layouts)).template(),
-        )
-        if all(layout.has_static_sizes for layout in layouts):
-            object.__setattr__(self, "_cached_anc_templates", templates)
-        return list(templates)
-
-    @property
-    def unitary(self) -> Callable[..., None]:
-        """Return the PREP-SELECT-PREP unitary derived from the terms.
-
-        The returned callable is built once and cached, and so are the SELECT
-        branches it dispatches to. Both matter for compilation cost: ``q_switch``
-        traces each branch several times per call, and re-deriving the branches on
-        every invocation would make each of those a full-price trace instead of a
-        cache hit.
-        """
-        cached = self.__dict__.get("_cached_unitary")
-        if cached is not None:
-            return cached
-
-        if len(self.terms) == 1:
-            coefficient, block_encoding = self.terms[0]
-            child_unitary = block_encoding.unitary
-
-            def unitary(*args):
-                child_unitary(*args)
-                with control(coefficient < 0):
-                    gphase(np.pi, args[0][0])
-
-            if not isinstance(coefficient, jax.core.Tracer) and _is_reusable_unitary(block_encoding):
-                object.__setattr__(self, "_cached_unitary", unitary)
-            return unitary
-
-        layouts = self._lcu_layouts
-        child_unitaries = [block_encoding.unitary for _, block_encoding in self.terms]
-
-        branches = [
-            _make_lcu_branch(child_unitary, layout, f"lcu_branch_{term_index}")
-            for term_index, (child_unitary, layout) in enumerate(zip(child_unitaries, layouts))
-        ]
-        # Padding to the full selector dimension also keeps the branch count even,
-        # which q_switch relies on: it appends an identity branch to an odd list,
-        # and must not be allowed to grow this cached one.
-        branches += [_identity_lcu_branch] * ((1 << self._lcu_selector_size) - len(branches))
-
-        # PREP acts on real amplitudes whenever the linear combination has no
-        # negative or complex coefficient, which lets the uncomputation be expressed
-        # as a conjugation instead of a separately traced inverse.
-        has_real_amplitudes = all(_is_non_negative_real(coefficient) for coefficient, _ in self.terms)
-
-        def unitary(*args):
-            selector = args[0]
-            shared_ancilla = args[1]
-            operands = args[2:]
-
-            amplitudes = self._lcu_amplitudes
-            xp = np if isinstance(amplitudes, np.ndarray) else jnp
-
-            if has_real_amplitudes:
-                with conjugate(prepare)(selector, xp.abs(amplitudes)):
-                    q_switch(selector, branches, shared_ancilla, *operands)
-            else:
-                prepare(selector, amplitudes)
-                q_switch(selector, branches, shared_ancilla, *operands)
-                with invert():
-                    prepare(selector, xp.conjugate(amplitudes))
-
-        cacheable = all(layout.has_static_sizes for layout in layouts) and all(
-            _is_reusable_unitary(block_encoding) for _, block_encoding in self.terms
-        )
-        if cacheable:
-            object.__setattr__(self, "_cached_unitary", unitary)
-        return unitary
-
-    @property
-    def num_ops(self) -> int:
-        """Return the common number of operands in the child encodings."""
-        return self.terms[0][1].num_ops
-
-    @property
-    def num_ancs(self) -> int:
-        """Return the number of selector and shared-workspace ancillas."""
-        return len(self._anc_templates)
-
-    @property
-    def is_hermitian(self) -> bool:
-        """Return whether every child encoding has a Hermitian unitary."""
-        return all(block_encoding.is_hermitian for _, block_encoding in self.terms)
-
-    def _get_lcu_terms(self) -> _LCUTerms:
-        return self.terms
-
-    def tree_flatten(self) -> tuple[tuple[ArrayLike | BlockEncoding, ...], int]:
-        """Flatten only the authoritative terms for JAX pytree handling."""
-        children = tuple(value for term in self.terms for value in term)
-        return children, len(self.terms)
-
-    @classmethod
-    def tree_unflatten(
-        cls: type[LinearCombinationBlockEncoding],
-        aux_data: int,
-        children: tuple[ArrayLike | BlockEncoding, ...],
-    ) -> LinearCombinationBlockEncoding:
-        """Reconstruct a linear combination from flattened terms."""
-        terms = tuple((children[index], children[index + 1]) for index in range(0, 2 * aux_data, 2))
-        return cls(terms)
